requirements.txt
aiogram~=3.16.0
python-dotenv~=1.0.1
torch~=2.5.1
torchvision~=0.20.1
pillow~=11.0.0
requests~=2.32.3